backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
supabase==2.28.0
python-dotenv==1.2.2
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.22
pillow==12.1.1
pydantic==2.12.5
httpx==0.28.1
bcrypt==4.3.0
folium>=0.15.0